jinglebytex-backend/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
colorama==0.4.6
databases==0.9.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
mysqlclient==2.2.6
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
